QuerySnListPacket的Request协议中版本号改成uint16_t类型
git-svn-id: http://svn01/repositories/r01/p2p/trunk@903 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ protocol/BootstrapPacket.h
@@ -443,7 +443,7 @@ namespace protocol
         // request
         QuerySnListPacket(
             boost::uint32_t transaction_id,
-            boost::uint32_t peer_version,
+            boost::uint16_t peer_version,
             boost::asio::ip::udp::endpoint endpoint_)
         {
             transaction_id_ = transaction_id;
